reorganize folder structure and have components dealing with pages in the /pages folder and fix div in LoadingBar.
