+++ src/components/LoadingBar/LoadingBar.jsx
@@ -4,26 +4,26 @@ const LoadingBar = () => {
   return (
     <div className="loading-bar">
       <div className="loading-bar__box-container">
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
-        <div className="loading-bar__box"></div>
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
+        <div className="loading-bar__box" />
       </div>
     </div>
   );
