+++ src/pages/LoadingPage/LoadingPage.scss
