+++ src/pages/QuizPage/QuizPage.scss
